requirements.txt
prefect==1.4.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
colorama==0.4.1
croniter==1.3.7
dask==2022.10.2
distributed==2022.10.2
fsspec==2022.10.0
greenlet==1.1.3.post0
HeapDict==1.0.1
idna==3.4
importlib-resources==5.10.0
Jinja2==3.1.2
locket==1.0.0
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-oneofschema==3.0.1
msgpack==1.0.4
mypy-extensions==0.4.3
numpy==1.23.4
packaging==21.3
pandas==1.5.1
partd==1.3.0
pendulum==2.1.2
psutil==5.9.3
pyodbc==4.0.34
pyparsing==3.0.9
python-box==6.1.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-slugify==6.1.2
pytz==2022.5
pytzdata==2020.1
pywin32==304
PyYAML==6.0
reactivex==4.0.4
requests==2.28.1
six==1.16.0
smartystreets-python-sdk==4.11.9
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.42
tabulate==0.9.0
tblib==1.7.0
text-unidecode==1.3
toml==0.10.2
toolz==0.12.0
tornado==6.1
typing_extensions==4.4.0
urllib3==1.26.12
websocket-client==1.4.1
zict==2.2.0
